+++ src/utils/CommandConfigBuilder.ts
@@ -47,6 +47,8 @@ export class CommandConfigBuilder {
     cmd.action = config.action;
     cmd.alias = config.alias || null;
     cmd.description = config.description || null;
+    cmd.usage = config.usage || null;
+    cmd.signature = config.signature || null;
     return cmd;
   }
 }